+++ arthas-react/index.js
@@ -1,0 +1,11 @@
+import {
+  PersonaForm,
+  PersonaChat,
+  PersonaList
+} from './src/components';
+
+export {
+  PersonaForm,
+  PersonaChat,
+  PersonaList
+};
