+++ arthas-react/src/App.css
@@ -4,89 +4,6 @@
   padding: 1rem;
 }
 
-.online-indicator {
-  display: block;
-  background: #999;
-  border-radius: 100vw;
-  min-width: .4rem;
-  min-height: .4rem;
-  max-width: .4rem;
-  max-height: .4rem;
-}
-
-.persona-list-item > .persona-avatar {
-  display: flex;
-  flex-direction: column;
-  min-width: 3rem;
-  min-height: 3rem;
-  max-width: 3rem;
-  max-height: 3rem;
-  filter: grayscale(1);
-  background: linear-gradient(to top right, violet, coral);
-  border-radius: 100vw;
-}
-
-.persona-avatar > img {
-  flex: 1;
-  border-radius: inherit;
-  border: 1px solid white;
-  object-fit: cover;
-  max-height: inherit;
-}
-
-.persona-list {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.persona-list-item {
-  display: flex;
-  gap: .5rem;
-  padding: .5rem 1rem;
-  align-items: center;
-  cursor: pointer;
-  opacity: .5;
-}
-
-.persona-list-item > h2 {
-  font-size: 1em;
-  font-weight: 400;
-  font-style: italic;
-}
-
-.persona-list-item.selected {
-  opacity: 1;
-}
-
-.persona-list-item.selected > .online-indicator {
-  background: #07f007;
-  box-shadow: 0 0 1rem #07f007;
-}
-
-.persona-list-item.selected > .persona-avatar {
-  filter: grayscale(0);
-}
-
-.persona-list-item.selected > h2 {
-  font-style: normal;
-  opacity: 1;
-}
-
-.persona-list-item > span {
-  display: block;
-}
-
-.form {
-  position: relative;
-  background: white;
-  color: #121212;
-  box-shadow: 0 0 10rem dodgerblue;
-  width: 75%;
-  padding: 1rem;
-  border-radius: 1rem;
-}
-
 #root {
   display: flex;
   gap: 1rem;
@@ -143,6 +60,11 @@
   opacity: 1;
 }
 
+#create-persona-button {
+  background: white;
+  color: dodgerblue;
+}
+
 #input,
 #output {
   display: flex;
@@ -155,7 +77,7 @@
   flex: 1;
 }
 
-#output .img {
+#output .img.full {
   width: calc(100% - 1rem);
   margin: 1rem .5rem;
   border-radius: 1rem;
@@ -163,7 +85,7 @@
   background: black;
 }
 
-#output .img > img {
+#output .img.full > img {
   border-radius: inherit;
 }
 
@@ -204,26 +126,9 @@
   color: dodgerblue;
 }
 
-#create-persona-button {
-  background: white;
-  color: dodgerblue;
-}
-
 @media (max-width: 900px) {
-  .persona-list-item {
-    padding: 0;
-    border-radius: 100vw;
-    margin: 0;
-  }
-
-  h1,
-  .persona-list-item > span,
-  .persona-list-item > h2 {
+  h1 {
     display: none;
-  }
-
-  .persona-list-item > .persona-avatar {
-    display: block;
   }
 
   .form {
