+++ arthas-react/src/components/PersonaChat/index.css
@@ -1,0 +1,60 @@
+#input,
+#output {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+#output {
+  flex: 1;
+}
+
+#output .img {
+  width: calc(100% - 1rem);
+  margin: 0 auto;
+  border-radius: 100vw;
+  box-shadow: 0 0 5rem black;
+  background: black;
+  width: 3rem;
+  height: 3rem;
+}
+
+#output .img > img {
+  border-radius: inherit;
+}
+
+#output p {
+  width: 60%;
+  margin: 1rem .5rem;
+  padding: 1rem;
+  background: dodgerblue;
+  border-radius: 1rem;
+  color: white;
+}
+
+#input > input,
+#input > button {
+  display: block;
+  min-height: 3rem;
+  margin-bottom: .5rem;
+  border: none;
+  border-radius: .5rem;
+  width: 100%;
+  appearance: none;
+  outline: none;
+  background: black;
+}
+
+#input > input {
+  background: #12121210;
+  padding: 0 .5rem;
+}
+
+#input > button {
+  background: dodgerblue;
+}
+
+#input > button:hover {
+  background-color: white;
+  color: dodgerblue;
+}
