+++ arthas-react/src/components/index.js
@@ -1,0 +1,9 @@
+import PersonaChat from './PersonaChat';
+import PersonaForm from './PersonaForm';
+import PersonaList from './PersonaList';
+
+export {
+  PersonaChat,
+  PersonaForm,
+  PersonaList
+};
