+++ examples/instant-assistant/instant-assistant/src/App.css
@@ -4,91 +4,7 @@
   padding: 1rem;
 }
 
-.online-indicator {
-  display: block;
-  background: #999;
-  border-radius: 100vw;
-  min-width: .4rem;
-  min-height: .4rem;
-  max-width: .4rem;
-  max-height: .4rem;
-}
-
-.persona-list-item > .persona-avatar {
-  display: flex;
-  flex-direction: column;
-  min-width: 3rem;
-  min-height: 3rem;
-  max-width: 3rem;
-  max-height: 3rem;
-  filter: grayscale(1);
-  background: linear-gradient(to top right, violet, coral);
-  border-radius: 100vw;
-}
-
-.persona-avatar > img {
-  flex: 1;
-  border-radius: inherit;
-  border: 1px solid white;
-  object-fit: cover;
-  max-height: inherit;
-}
-
-.persona-list {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.persona-list-item {
-  display: flex;
-  gap: .5rem;
-  padding: .5rem 1rem;
-  align-items: center;
-  cursor: pointer;
-  opacity: .5;
-}
-
-.persona-list-item > h2 {
-  font-size: 1em;
-  font-weight: 400;
-  font-style: italic;
-}
-
-.persona-list-item.selected {
-  opacity: 1;
-}
-
-.persona-list-item.selected > .online-indicator {
-  background: #07f007;
-  box-shadow: 0 0 1rem #07f007;
-}
-
-.persona-list-item.selected > .persona-avatar {
-  filter: grayscale(0);
-}
-
-.persona-list-item.selected > h2 {
-  font-style: normal;
-  opacity: 1;
-}
-
-.persona-list-item > span {
-  display: block;
-}
-
-.form {
-  position: relative;
-  background: white;
-  color: #121212;
-  box-shadow: 0 0 10rem dodgerblue;
-  width: 75%;
-  padding: 1rem;
-  border-radius: 1rem;
-}
-
-#app,
-#nav {
+#app {
   display: flex;
   flex-direction: column;
   flex: 1;
@@ -107,10 +23,6 @@
   right: 1rem;
   bottom: 6rem;
   z-index: 1000;
-}
-
-#nav {
-  display: none;
 }
 
 #overlay {
@@ -143,67 +55,6 @@
 #overlay.show > .form {
   margin-top: 0;
   opacity: 1;
-}
-
-#input,
-#output {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-#output {
-  flex: 1;
-}
-
-#output .img {
-  width: calc(100% - 1rem);
-  margin: 0 auto;
-  border-radius: 100vw;
-  box-shadow: 0 0 5rem black;
-  background: black;
-  width: 3rem;
-  height: 3rem;
-}
-
-#output .img > img {
-  border-radius: inherit;
-}
-
-#output p {
-  width: 60%;
-  margin: 1rem .5rem;
-  padding: 1rem;
-  background: dodgerblue;
-  border-radius: 1rem;
-  color: white;
-}
-
-#input > input,
-#input > button {
-  display: block;
-  min-height: 3rem;
-  margin-bottom: .5rem;
-  border: none;
-  border-radius: .5rem;
-  width: 100%;
-  appearance: none;
-  outline: none;
-  background: black;
-}
-
-#input > input {
-  background: #12121210;
-  padding: 0 .5rem;
-}
-
-#input > button {
-  background: dodgerblue;
-}
-
-#input > button:hover {
-  background-color: white;
-  color: dodgerblue;
 }
 
 #create-persona-button {
