+++ examples/instant-assistant/instant-assistant/src/App.js
@@ -1,27 +1,15 @@
 import { useEffect, useState } from 'react';
 
-import RandomProfile from 'random-profile-generator';
+import {
+  PersonaChat,
+  PersonaForm
+} from 'arthas-react';
 
 import './App.css';
 
-const BASE_URL = 'http://localhost:8000';
 const STORAGE_KEY = 'ARTHAS_INSTANT_ASSISTANT';
 const OVERLAY = 'overlay';
-const SEND = 'Send';
-const SAVE = 'Save';
 const CREATE = '🗨';
-const PERSONA_LIST_HEADING = 'Domain-specific Personas (DSPs)';
-const PERSONA_CREATED = 'New persona created.';
-const PERSONA_ERROR = 'Error loading persona.';
-const PERSONA_NAME = 'Name';
-const PERSONA_AVATAR_URL = 'Avatar URL';
-const PERSONA_KNOWLEDGE_URI = 'Store URL';
-const PERSONA_WRITING_STYLE = 'Writing style';
-const PERSONA_WRITING_TONE = 'Writing tone';
-// const DEFAULT_PERSONA_GENDER = 'Male';
-const DEFAULT_PERSONA_GENDER = 'Female';
-const SAVE_ERROR = 'Error saving persona.';
-const API_ERROR = 'API temporarily unavailable.';
 
 const SAVED_PERSONAS = JSON.parse(
   localStorage.getItem(STORAGE_KEY)
@@ -42,7 +30,6 @@ const App = () => {
   const [personaWritingStyle, setPersonaWritingStyle] = useState('');
   const [personaWritingTone, setPersonaWritingTone] = useState('');
   const [personaAvatarURL, setPersonaAvatarURL] = useState('');
-  const [personaGender] = useState(DEFAULT_PERSONA_GENDER);
 
   useEffect(() => {
     const personas = getPersonasArray();
@@ -67,11 +54,23 @@ const App = () => {
   }, []);
 
   useEffect(() => {
-    document.body.onkeydown = isCreating
-      ? (overlayClassName && onKeyDownOverlay)
-      : null;
+    const bindKeyDown = () => (
+      document.body.onkeydown = (overlayClassName && onKeyDownOverlay)
+    );
 
-    return () => document.body.onkeydown = null;
+    const clearKeyDown = () => (
+      document.body.onkeydown = null
+    );
+
+    if (isCreating) {
+      bindKeyDown();
+      setDisabled(true);
+    } else {
+      clearKeyDown();
+      setDisabled(false);
+    }
+
+    return () => clearKeyDown();
 
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isCreating, overlayClassName]);
@@ -80,65 +79,16 @@ const App = () => {
     return () => clearTimeout(timeoutId);
   }, [timeoutId]);
 
-  const ask = async () => {
-    setDisabled(true);
-
-    const response = await fetch(`${BASE_URL}/v1/prompt`, {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        key: persona.knowledgeURI,
-        input: question
-      })
-    });
-
-    if (response?.ok) {
-      const { error, answer = {} } = await response.json();
-
-      if (answer.pending) {
-        window.location.reload();
-
-        return;
-      }
-
-      if (error) {
-        alert(error.message || API_ERROR);
-        setDisabled(false);
-
-        return;
-      } else {
-        setText(answer.text);
-      }
-    }
-
-    setQuestion('');
-    setDisabled(false);
-  };
-
   const openOverlay = () => {
-    const {
-      firstName,
-      avatar
-    } = RandomProfile.profile(personaGender);
-
     setIsCreating(true);
     setOverlayClassName('');
 
     requestAnimationFrame(() => {
-      setPersonaName(firstName);
-      setPersonaAvatarURL(avatar);
       setOverlayClassName('show');
     });
   };
 
   const closeOverlay = () => {
-    setPersonaName('');
-    setPersonaKnowledgeURI('');
-    setPersonaWritingStyle('');
-    setPersonaWritingTone('');
     setOverlayClassName('');
 
     setTimeoutId(
@@ -148,20 +98,8 @@ const App = () => {
     );
   };
 
-  const getQuestionPlaceholder = () => (
-    !persona ? '...' : `What would you like to ask ${persona.name}?`
-  );
-
   const getPersonasArray = () => (
     Object.values(personaList || {})
-  );
-
-  const onChangeQuestion = ({ target: { value }}) => {
-    setQuestion(value);
-  };
-
-  const onKeyDownQuestion = ({ keyCode }) => (
-    question && keyCode === 13 && ask()
   );
 
   const onKeyDownOverlay = ({ keyCode }) => {
@@ -170,114 +108,10 @@ const App = () => {
     }
   };
 
-  const onClickCreate = openOverlay;
-
-  const onClickSave = async () => {
-    setDisabled(true);
-
-    const personaConfig = {
-      name: personaName,
-      knowledgeURI: personaKnowledgeURI,
-      avatarURL: personaAvatarURL,
-
-      // Passing null for artStyle will skip image generation
-
-      artStyle: null,
-      writingStyle: personaWritingStyle,
-      writingTone: personaWritingTone
-    };
-
-    const updatedCurrentPersona = { ...persona };
-
-    updatedCurrentPersona.online = false;
-
-    const updatedPersonaList = {
-      ...personaList,
-
-      [updatedCurrentPersona.knowledgeURI]: updatedCurrentPersona,
-      [personaKnowledgeURI]: personaConfig
-    };
-
-    const response = await fetch(`${BASE_URL}/v1/configure`, {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify(personaConfig)
-    });
-
-    if (response?.ok) {
-      const { error, success } = await response.json();
-
-      if (success) {
-        localStorage.setItem(
-          STORAGE_KEY,
-          JSON.stringify(updatedPersonaList)
-        );
-
-        alert(PERSONA_CREATED);
-        window.location.reload();
-
-        return;
-      } else {
-        alert(error?.message || SAVE_ERROR);
-      }
-    } else {
-      alert(SAVE_ERROR);
-    }
-
-    setDisabled(false);
-  };
-
   const onClickOverlay = ({ target: { id }}) => {
     if (id === OVERLAY) {
       closeOverlay();
     }
-  };
-
-  const onClickPersonaListItem = key => async () => {
-    setDisabled(true);
-
-    const updatedCurrentPersona = { ...persona };
-
-    updatedCurrentPersona.online = false;
-
-    const selectedPersona = {
-      ...personaList[key],
-
-      online: true
-    };
-
-    setPersona(selectedPersona);
-
-    setPersonaList({
-      ...personaList,
-
-      [updatedCurrentPersona.knowledgeURI]: updatedCurrentPersona,
-      [selectedPersona.knowledgeURI]: selectedPersona
-    });
-
-    const response = await fetch(`${BASE_URL}/v1/configure`, {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify(selectedPersona)
-    });
-
-    if (response?.ok) {
-      const { success } = await response.json();
-
-      if (!success) {
-        alert(PERSONA_ERROR);
-      }
-    }
-
-    setText('');
-    setQuestion('');
-    setDisabled(false);
   };
 
   const onChangePersonaName = ({ target: { value }}) => (
@@ -300,116 +134,51 @@ const App = () => {
     setPersonaAvatarURL(value)
   );
 
+  const onQuestion = value => setQuestion(value);
+
+  const onAnswer = answer => setText(answer?.text);
+
+  const personaFormProps = {
+    textOnly: true,
+    disabled,
+    persona,
+    personaList,
+    personaName,
+    personaKnowledgeURI,
+    personaArtStyle,
+    personaWritingStyle,
+    personaWritingTone,
+    personaAvatarURL,
+    onChangePersonaName,
+    onChangePersonaKnowledgeURI,
+    onChangePersonaWritingStyle,
+    onChangePersonaWritingTone,
+    onChangePersonaAvatarURL
+  };
+
+  const personaChatProps = {
+    disabled: disabled || isCreating,
+    persona,
+    question,
+    image,
+    text,
+    onQuestion,
+    onAnswer
+  };
+
   return <>
     {isCreating && <aside id="overlay" className={overlayClassName} onClick={onClickOverlay}>
-      <div className="form">
-        <h2>Create Shopping Assistant</h2>
-        <input
-          required
-          disabled={disabled}
-          placeholder={PERSONA_NAME}
-          onChange={onChangePersonaName}
-          value={personaName}
-        />
-        <input
-          required
-          disabled={disabled}
-          placeholder={PERSONA_AVATAR_URL}
-          onChange={onChangePersonaAvatarURL}
-          value={personaAvatarURL}
-        />
-        <input
-          required
-          disabled={disabled}
-          placeholder={PERSONA_KNOWLEDGE_URI}
-          onChange={onChangePersonaKnowledgeURI}
-          value={personaKnowledgeURI}
-        />
-        <input
-          required
-          disabled={disabled}
-          placeholder={PERSONA_WRITING_STYLE}
-          onChange={onChangePersonaWritingStyle}
-          value={personaWritingStyle}
-        />
-        <input
-          required
-          disabled={disabled}
-          placeholder={PERSONA_WRITING_TONE}
-          onChange={onChangePersonaWritingTone}
-          value={personaWritingTone}
-        />
-        <button
-          disabled={disabled}
-          onClick={onClickSave}
-        >
-          {SAVE}
-        </button>
-      </div>
+      <PersonaForm {...personaFormProps} />
     </aside>}
-    <nav id="nav" className="panel">
-      <h1>{PERSONA_LIST_HEADING}</h1>
-      <ul className="persona-list">
-        {personaList && getPersonasArray().map(({
-          name,
-          knowledgeURI,
-          avatarURL,
-          online = false
-        }) => online && (
-          <li
-            key={knowledgeURI}
-            className={`persona-list-item panel ${online ? 'selected' : ''}`}
-            onClick={onClickPersonaListItem(knowledgeURI)}
-          >
-            <span className="online-indicator" />
-            <span className="persona-avatar">
-              {avatarURL && <img
-                src={avatarURL}
-                alt={name}
-                width="100%"
-                height="100%"
-              />}
-            </span>
-            <h2>{name}</h2>
-          </li>
-        ))}
-      </ul>
-    </nav>
     <button
       disabled={isCreating}
       id="create-persona-button"
-      onClick={onClickCreate}
+      onClick={openOverlay}
     >
       {CREATE}
     </button>
     {persona && <main id="app" className="panel">
-      <div id="output">
-        <div className="img">
-          {persona.avatarURL && <img
-            src={persona.avatarURL}
-            alt={persona.name}
-            width="100%"
-            height="100%"
-          />}
-        </div>
-        {text && <p>{text}</p>}
-      </div>
-      <div id="input">
-        <input
-          autoFocus
-          disabled={disabled || isCreating}
-          value={question}
-          placeholder={getQuestionPlaceholder()}
-          onChange={onChangeQuestion}
-          onKeyDown={onKeyDownQuestion}
-        />
-        <button
-          disabled={disabled || isCreating}
-          onClick={ask}
-        >
-          {SEND}
-        </button>
-      </div>
+      <PersonaChat {...personaChatProps } />
     </main>}
   </>;
 }
